+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## HEAD
 
+* `file_system` – Deprecate `:older_than` option for `FileSystem#clear!` in favour of a block (@janko)
+
+* `file_system` – Add the ability for `FileSystem#clear!` to take a block (@janko)
+
 * `signature` – Add `Shrine.signature` alias for `Shrine.calculcate_signature` (@janko)
 
 * `store_dimensions` – Add `Shrine.dimensions` alias for `Shrine.extract_dimensions` (@janko)